visionU/requirements.txt
Flask==3.0.0
pandas==2.1.4
reportlab==4.0.9
PyMuPDF==1.23.8
PyPDF2==3.0.1
openpyxl==3.1.2